+++ backend/src/main/scala/jooq/db/tables/daos/TextaudioindexDao.scala
@@ -4,7 +4,6 @@
 package jooq.db.tables.daos
 
 
-import java.lang.Byte
 import java.lang.Double
 import java.lang.Integer
 import java.util.List
@@ -58,7 +57,7 @@ class TextaudioindexDao(configuration : Configuration) extends DAOImpl[Textaudio
     fetch(Textaudioindex.TEXTAUDIOINDEX.SPEAKERKEY, values:_*)
   }
 
-  def fetchByLabeled(values : Byte*) : List[jooq.db.tables.pojos.Textaudioindex] = {
+  def fetchByLabeled(values : Integer*) : List[jooq.db.tables.pojos.Textaudioindex] = {
     fetch(Textaudioindex.TEXTAUDIOINDEX.LABELED, values:_*)
   }
